feat(riot): separate tests
