+++ riot/status_test.go
@@ -1,0 +1,69 @@
+package riot
+
+import (
+	"fmt"
+	"net/http"
+	"testing"
+
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/KnutZuidema/golio/api"
+	"github.com/KnutZuidema/golio/internal"
+	"github.com/KnutZuidema/golio/internal/mock"
+)
+
+func TestStatusClient_Get(t *testing.T) {
+	t.Parallel()
+	tests := []struct {
+		name    string
+		want    *Status
+		doer    internal.Doer
+		wantErr error
+	}{
+		{
+			name: "get response",
+			want: &Status{},
+			doer: mock.NewJSONMockDoer(Status{}, 200),
+		},
+		{
+			name: "unknown error status",
+			wantErr: api.Error{
+				Message:    "unknown error reason",
+				StatusCode: 999,
+			},
+			doer: mock.NewStatusMockDoer(999),
+		},
+		{
+			name:    "not found",
+			wantErr: api.ErrNotFound,
+			doer:    mock.NewStatusMockDoer(http.StatusNotFound),
+		},
+		{
+			name: "rate limited",
+			want: &Status{},
+			doer: rateLimitDoer(Status{}),
+		},
+		{
+			name: "unavailable once",
+			want: &Status{},
+			doer: unavailableOnceDoer(Status{}),
+		},
+		{
+			name:    "unavailable twice",
+			wantErr: api.ErrServiceUnavailable,
+			doer:    mock.NewStatusMockDoer(http.StatusServiceUnavailable),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			client := NewClient(api.RegionEuropeWest, "API_KEY", tt.doer, logrus.StandardLogger())
+			got, err := client.Status.Get()
+			require.Equal(t, err, tt.wantErr, fmt.Sprintf("want err %v, got %v", tt.wantErr, err))
+			if tt.wantErr == nil {
+				assert.Equal(t, got, tt.want)
+			}
+		})
+	}
+}
